requirements.txt
BeautifulSoup==3.2.1
beautifulsoup4==4.4.1
bs4==0.0.1
flake8==2.5.2
mccabe==0.4.0
pep8==1.7.0
pyflakes==1.0.0
python-memcached==1.57
requests==2.9.1
six==1.10.0
wsgiref==0.1.2
